Updated Scrpple to V3 and organization
I imported the relevant Scrpple V3 variables and maps from the GA subproject. Runs for the whole landscape. This adds in the new accounting of mortality and spread, that seem to capture the fire regime for the GA subset. Suppression effectiveness also upgraded but should be checked against Southern Apps data. I also organized alot which accounts for most of the changes you see in the log. If you are missing somehting it should now be in "Bonus_Tracks" or a file corresponding to the extension. Fine fuels likely still aproblem, so they are removed from spread.